requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
google-ads==28.4.1
    # via -r requirements/base.in
google-api-core[grpc]==2.29.0
    # via
    #   google-ads
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.47.0
    # via
    #   -r requirements/base.in
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-storage
google-auth-oauthlib==1.2.4
    # via
    #   -r requirements/base.in
    #   google-ads
google-cloud-bigquery==3.40.0
    # via -r requirements/base.in
google-cloud-bigquery-storage==2.36.0
    # via -r requirements/base.in
google-cloud-core==2.5.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-storage==3.8.0
    # via -r requirements/base.in
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.72.0
    # via
    #   google-ads
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-ads
    #   google-api-core
    #   google-cloud-bigquery-storage
    #   grpcio-status
grpcio-status==1.76.0
    # via
    #   google-ads
    #   google-api-core
idna==3.11
    # via requests
numpy==2.4.1
    # via pandas
oauthlib==3.3.1
    # via requests-oauthlib
orjson==3.11.5
    # via -r requirements/base.in
packaging==25.0
    # via google-cloud-bigquery
pandas==2.3.3
    # via -r requirements/base.in
proto-plus==1.27.0
    # via
    #   google-ads
    #   google-api-core
    #   google-cloud-bigquery-storage
protobuf==6.33.4
    # via
    #   google-ads
    #   google-api-core
    #   google-cloud-bigquery-storage
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyarrow==22.0.0
    # via -r requirements/base.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.in
    #   google-cloud-bigquery
    #   pandas
python-dotenv==1.2.1
    # via -r requirements/base.in
pytz==2025.2
    # via
    #   -r requirements/base.in
    #   pandas
pyyaml==6.0.3
    # via google-ads
requests==2.32.5
    # via
    #   -r requirements/base.in
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
tenacity==9.1.2
    # via -r requirements/base.in
typing-extensions==4.15.0
    # via grpcio
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
